requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
cloudinary==1.40.0
contourpy==1.2.0
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.51.0
gitdb==4.0.7
GitPython==3.1.37
idna==3.7
Jinja2==3.1.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.0
mplsoccer==1.2.4
numexpr==2.8.7
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pip==24.0
protobuf==3.20.3
pyarrow==14.0.2
pydeck==0.8.0
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2024.1
referencing==0.30.2
requests==2.31.0
rich==13.3.5
rpds-py==0.10.6
scipy==1.13.0
seaborn==0.12.2
setuptools==69.5.1
six==1.16.0
smmap==4.0.0
streamlit==1.32.0
streamlit-dynamic-filters==0.1.6
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.11.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.1.0
wheel==0.43.0
